[feat] - 로직별 컴포넌트 분리
