+++ app/components/activityRecommend/DetailedJobSelection.tsx
@@ -1,0 +1,7 @@
+// 세부 직무 선택
+
+const DetailedJobSelection = () => {
+  return <div>DetailedJobSelection</div>;
+};
+
+export default DetailedJobSelection;
