+++ app/components/activityRecommend/JobSelection.tsx
@@ -1,0 +1,19 @@
+// 직무 선택 (직무 선택 드롭다운)
+const JobSelection = () => {
+  return (
+    <>
+      <select>
+        <option value={0}>웹개발</option>
+        <option value={1}>백엔드 개발자</option>
+        <option value={2}>23</option>
+      </select>
+      <select>
+        <option value={1}>백엔드 개발자</option>
+        <option value={2}>23</option>
+        <option value={3}>23</option>
+      </select>
+    </>
+  );
+};
+
+export default JobSelection;
