+++ app/components/activityRecommend/PageNation.tsx
@@ -1,0 +1,6 @@
+// 페이지네이션
+const PageNation = () => {
+  return <div>PageNation</div>;
+};
+
+export default PageNation;
