+++ app/types/field-types.ts
@@ -1,0 +1,4 @@
+export interface Field {
+  title: string;
+  imageUrl: string;
+}
